requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.4.2 